Some date stuff

+++ src/java/no/hib/mod250/enterpriseBeans/ProductDAO.java
@@ -48,12 +48,12 @@ public class ProductDAO {
         
     }
     
-    public List<Product> getProductsByUser(int userId) {
+    public List<Product> getProductsByUser(long userId) {
         Query query = em.createQuery("SELECT u FROM Product u WHERE u.sellerId = :userId");
         return query.setParameter("userId", userId).getResultList();
     }
     
-    public Product getProductById(int productId) {
+    public Product getProductById(long productId) {
         Query query = em.createQuery("SELECT u FROM Product u WHERE u.id = :productId");
         return (Product) query.setParameter("productId", productId).getResultList().get(0);
     }
